Introduce context-based evaluation for abbreviation replacement
There is ambiguity in replacing abbreviations. In practice, an abbreviation
can have multiple meanings. For example, "ck" can mean "chồng" or "chung kết"
causing difficulties in restoring the original word. Based on the observation
that selecting a replacement for an abbreviation depends on the context of
the text: "vòng ck" -> "vòng chung kết" and "vk ck son" -> "vợ chồng son".
This commit introduces context-based evaluation for abbreviation replacement
using a simple trigram model to calculate the likelihood of a replacement
word given the context. The replacement word with the highest probability is
then selected.

Signed-off-by: hoang1007 <hoangvuhuy1007@gmail.com>